renames localized-v-tooltip in tooltip
