+++ modules/@apostrophecms/ui/ui/apos/lib/vue.js
@@ -1,6 +1,6 @@
 import { createApp } from 'vue';
 import ClickOutsideElement from './click-outside-element';
-import LocalizedVTooltip from './localized-v-tooltip';
+import Tooltip from './tooltip';
 import VueAposI18Next from './i18next';
 
 export default (appConfig, props = {}) => {
@@ -10,7 +10,7 @@ export default (appConfig, props = {}) => {
     // Module aliases are not available yet when this code executes
     i18n: apos.modules['@apostrophecms/i18n']
   });
-  app.use(LocalizedVTooltip);
+  app.use(Tooltip);
   app.use(ClickOutsideElement);
 
   window.apos.registerVueComponents(app);
